+++ src/Exception/FeatureNotSupportedException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Sokil\Mongo\Exception;
+
+class FeatureNotSupportedException extends \Sokil\Mongo\Exception
+{
+
+}