+++ tests/CollectionTest.php
@@ -1571,6 +1571,22 @@ class CollectionTest extends \PHPUnit_Framework_TestCase
         ), $indexes[1]['key']);
 
     }
+    
+    public function testDeleteIndex()
+    {
+        $this->collection->ensureIndex(array(
+            'asc'    => 1,
+            'desc'   => -1,
+        ));
+
+        $this->collection->deleteIndex(array(
+            'asc'    => 1,
+            'desc'   => -1,
+        ));
+        
+        $indexes = $this->collection->getIndexes();
+        $this->assertEquals(1, count($indexes));
+    }
 
     public function testEnsureSparseIndex()
     {
@@ -1679,14 +1695,19 @@ class CollectionTest extends \PHPUnit_Framework_TestCase
 
     public function testEnsureFulltextIndex()
     {
-        $this->collection->ensureFulltextIndex(
-            array('fieldname1', 'fieldname2'),
-            array(
-                'fieldname1' => 1,
-                'fieldname2' => 2,
-            ),
-            'spanish'
-        );
+        try {
+            $this->collection->ensureFulltextIndex(
+                array('fieldname1', 'fieldname2'),
+                array(
+                    'fieldname1' => 1,
+                    'fieldname2' => 2,
+                ),
+                'spanish'
+            );
+        } catch (\MongoWriteConcernException $e) {
+            $this->assertEquals('127.0.0.1:27017: text search not enabled', $e->getMessage());
+            return;
+        }
 
         $indexes = $this->collection->getIndexes();
         $index = $indexes[1];